Add project.rootdir that always points to toplevel
This is a bit of a burden to maintain when projects are added or moved,
but makes it easier to point to files off the root path. Usually doing
so is a bad idea anyway, but in some cases it is nice to have access to
shared configuration, such as the files in the config folder.

This used to be revision r5471.